Split topk program factories
