+++ ttnn/cpp/ttnn/operations/reduction/topk/device/topk_device_operation.hpp
@@ -11,7 +11,8 @@
 #include "ttnn/decorators.hpp"
 
 #include "ttnn/operations/reduction/topk/device/topk_device_operation_types.hpp"
-#include "ttnn/operations/reduction/topk/device/topk_program_factory.hpp"
+#include "ttnn/operations/reduction/topk/device/topk_single_core_program_factory.hpp"
+#include "ttnn/operations/reduction/topk/device/topk_multi_core_program_factory.hpp"
 
 namespace ttnn::operations::reduction::topk {
