+++ ttnn/cpp/ttnn/operations/reduction/topk/device/topk_single_core_program_factory.hpp
@@ -4,8 +4,6 @@
 
 #pragma once
 
-#include <vector>
-
 #include "ttnn/device_operation.hpp"
 #include "ttnn/operations/reduction/topk/device/topk_device_operation_types.hpp"
 
@@ -19,29 +17,6 @@ struct TopKSingleCoreSharedVariables {
 
 struct TopKSingleCoreProgramFactory {
     using shared_variables_t = TopKSingleCoreSharedVariables;
-    using cached_program_t = ttnn::device_operation::CachedProgram<shared_variables_t>;
-
-    static cached_program_t create(
-        const operation_attributes_t& operation_attributes,
-        const tensor_args_t& tensor_args,
-        tensor_return_value_t& tensor_return_value);
-
-    static void override_runtime_arguments(
-        cached_program_t& cached_program,
-        const operation_attributes_t& operation_attributes,
-        const tensor_args_t& tensor_args,
-        tensor_return_value_t& tensor_return_value);
-};
-
-struct TopKMultiCoreSharedVariables {
-    tt::tt_metal::KernelHandle unary_reader_kernel_id{};
-    tt::tt_metal::KernelHandle binary_writer_final_kernel_id{};
-    std::vector<tt::tt_metal::CoreCoord> local_cores;
-    tt::tt_metal::CoreCoord final_core;
-};
-
-struct TopKMultiCoreProgramFactory {
-    using shared_variables_t = TopKMultiCoreSharedVariables;
     using cached_program_t = ttnn::device_operation::CachedProgram<shared_variables_t>;
 
     static cached_program_t create(
